+++ src/overrides/etc.js
@@ -1,6 +1,0 @@
-/** @type { import('eslint').Linter.RulesRecord } */
-export const etcRules = {
-  'etc/no-const-enum': 'error',
-  'etc/no-enum': 'error',
-  'etc/no-t': 'error',
-}