+++ src/overrides/index.js
@@ -1,5 +1,4 @@
 import { eslintRules } from './eslint.js'
-import { etcRules } from './etc.js'
 import { importRules } from './import.js'
 import { prettierRules } from './prettier.js'
 import { promiseRules } from './promise.js'
@@ -10,7 +9,6 @@ import { unicornRules } from './unicorn.js'
 /** @type { import('eslint').Linter.RulesRecord } */
 export const overrides = {
   ...eslintRules,
-  ...etcRules,
   ...importRules,
   ...prettierRules,
   ...promiseRules,
@@ -21,10 +19,4 @@ export const overrides = {
 
 export const overridesWithTypeChecking = {
   ...overrides,
-
-  // temporarily disabled since eslint-etc-plugin does not support flat config fully
-  // see https://github.com/import-js/eslint-plugin-import/pull/2714
-  'etc/no-commented-out-code': 'off',
-  'etc/no-misused-generics': 'error',
-  'etc/throw-error': 'error',
 }